automated_omr_full/requirements.txt
streamlit
opencv-python-headless
numpy
scikit-image
pillow
pandas
sqlalchemy
openpyxl